requirements.txt
ddapi==0.4.9
pydantic
python-dotenv
aiohttp
psycopg2-binary
